Render document metadata
The taxon facet does not render metadata, so adjusting the metadata rendering logic to work for a nested facet has its challenges, since the pattern does not exist in the codebase.

A nested facet, or a taxon facet, whilst conceptually a unit, must cater to two sets of data corresponding to the two underlying "facets". The "raw facets" we get from content store do not have this distinction, as there is only one facet there with a nested structure. So there needs to be some digging, mapping and flattening to compose the metadata. These raw facets are passed around a lot, making it difficult to isolate the change.

The metadata itself has a notion of "type", which in the pre-nested-facet world was either "text" or "date" and it was deduced from the type of the facet itself. I managed to implement a metadata "nested" type all the way through, causing quite a bit of changes in the process. Ultimately, I reverted that approach in favour of the current commit's implementation, because it seems to me the metadata type is for internal use - ultimately the distinction is whether to render a date or a text. Nested facets are essentially text and their metadata suffices to be just text.

I also replicated some of the tests to cover the nested facet case.