Load js for nested facets
Since the view now includes both selects, and the facet is unified, we can load the JS at the top level without any additional "is main facet?" checks, as was the case before.

This now enables us to keep some of the query selectors scoped to the module rather than the document.

Note: there is definitely scope here to bring the two js files together (for taxon and nested facets) if further refactoring is desired.