Add query_params to new NestedFacet
This method seems to have been added in c86175d7 for all facets.

I noticed no errors and no changes with or without this method defined. I could not locate any relevant uses of it either. We already define some query_params as part of the applied filters, perhaps that suffices.

Adding this for parity with other facets.