Let Search API v2 decide default serving config
We currently define a "default" serving config (when neither an AB test
nor a debug flag needs to override it) both in this app and in Search
API v2 itself.

This makes it harder to update what the default is as we'd have to do
it in two places at the same time.

Instead, this changes Finder Frontend to just not specify a serving
config in situations where we don't want to override the default, and
leaves it to Search API v2 to figure out.