Update finder content item documentation to reflect sorting validation rules
It does look like Finder Frontend requires having `relevance` sort option by default. When sort options are supplied without it, Javascript throws an error in the DOM as it tries to query the option on load. Consequently, a user will be unable to perform other JS interactions such as realtime filtering, sorting, etc.

This commit highlights the importance of having the required sort field as part of the options.