Temporarily use a different non-live paths in tests
These tests broke because the test was loading the real content item rather than the test fixture.

Tweaking the path to be something that is not live, means that the tests can run as intended.

This might be something we want to adopt in future, as it'll help prevent us from coupling behaviour to a path which might change.