Fix cucumber specs to work with offline mode
The cucumber specs are tied into real paths, which we're actually using as part of the offline mode.

So to make the specs work, I've tweaked most of the stubbed content items to use paths with a "-temp" suffix, and swapped out the fake taxons for real ones.

I've had to change some of the search stubs too, because the default filters have changed in some cases. It's quite hard to unravel some of these!